+++ docs/changelog.rst
@@ -31,6 +31,7 @@ Fixes
 * :class:`aspecd.processing.Normalisation` removes unit from last axis
 * :class:`aspecd.processing.BaselineCorrection` handles zero values in range properly
 * :class:`aspecd.analysis.AggregateAnalysisStep` no longer adds ``datasets`` and ``result`` to output of ``to_dict()``
+* :class:`aspecd.tasks.AggregatedAnalysisTask` sets correct type in output of ``to_dict()``
 
 
 Version 0.6.2